+++ src/index.js
@@ -1,8 +1,6 @@
-import React, { StrictMode } from "react";
+import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-
-import "./languages/i18n";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));