+++ src/pages/AMLPolicy.jsx
@@ -1,16 +1,14 @@
 import AMLPolicyBody from "../components/amlpolicyBody/amlpolicyBody";
 import { Helmet } from "react-helmet";
-import { useTranslation } from "react-i18next";
 import NefentusLogo from "../assets/logo/logo.svg";
-import { Conclusion } from "./landing";
-import "./landing/landing.css";
+import { Conclusion } from "../components/landing";
+import "../components/landing/landing.css";
 
 const AMLPolicy = () => {
-  const { t } = useTranslation();
   return (
     <div className="landing-layout container">
       <Helmet>
-        <title>Nefentus | {t("navigation.amlpolicy")}</title>
+        <title>Nefentus | AML policy</title>
       </Helmet>
       <AMLPolicyBody />
       <Conclusion