+++ src/pages/BusinessSupport.jsx
@@ -1,0 +1,15 @@
+import { Helmet } from "react-helmet";
+import BusinessSupportBody from "../components/landing/BusinessSupport";
+
+const BusinessSupport = () => {
+  return (
+    <div>
+      <Helmet>
+        <title>Nefentus | Contact</title>
+      </Helmet>
+      <BusinessSupportBody />
+    </div>
+  );
+};
+
+export default BusinessSupport;