+++ src/pages/Imprint.jsx
@@ -1,13 +1,11 @@
-import { useTranslation } from "react-i18next";
 import ImprintBody from "../components/imprintBody/imprintBody";
 import { Helmet } from "react-helmet";
 
 const Imprint = () => {
-  const { t } = useTranslation();
   return (
     <div>
       <Helmet>
-        <title>Nefentus | {t("navigation.imprint")}</title>
+        <title>Nefentus | Imprint</title>
       </Helmet>
       <ImprintBody />
     </div>