+++ src/pages/TechnicalSupport.jsx
@@ -1,0 +1,15 @@
+import { Helmet } from "react-helmet";
+import TechnicalSupportBody from "../components/landing/TechnicalSupport";
+
+const TechnicalSupport = () => {
+  return (
+    <div>
+      <Helmet>
+        <title>Nefentus | Contact</title>
+      </Helmet>
+      <TechnicalSupportBody />
+    </div>
+  );
+};
+
+export default TechnicalSupport;