+++ src/pages/Vacancy.jsx
@@ -1,13 +1,11 @@
-import { useTranslation } from "react-i18next";
 import { Helmet } from "react-helmet";
 import VacancyBody from "../components/vacancyBody";
 
 const Vacancy = () => {
-  const { t } = useTranslation();
   return (
     <div>
       <Helmet>
-        <title>Nefentus | {t("navigation.vacancy")}</title>
+        <title>Nefentus | Vacancy</title>
       </Helmet>
       <VacancyBody />
     </div>