+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # New changes
 
 - Add new blocks (audience, api) && Update some texts (MINOR)
+- Make landing separate (MAJOR)
 - Add new policy pages && add new blocks in landing (MINOR)
 - Update resource navbar && hero && headings of home page (PATCH)
 - Update product invoice work flow (MINOR)